+++ packages/web-component/src/lib/constants/index.ts
@@ -17,6 +17,7 @@ export const URL_REDIRECT_AUTH_CALLBACK_PARAM_NAME = 'ra-callback';
 export const URL_REDIRECT_AUTH_INITIATOR_PARAM_NAME = 'ra-initiator';
 export const DESCOPE_ATTRIBUTE_PREFIX = 'data-descope-';
 export const DESCOPE_ATTRIBUTE_EXCLUDE_FIELD = 'data-exclude-field';
+export const DESCOPE_ATTRIBUTE_EXCLUDE_NEXT_BUTTON = 'data-exclude-next';
 export const DESCOPE_LAST_AUTH_LOCAL_STORAGE_KEY = 'dls_last_auth';
 export const OIDC_IDP_STATE_ID_PARAM_NAME = 'state_id';